wip: detect flawed opening insertion
