+++ jolt-atlas-core/src/onnx_proof/prover.rs
@@ -28,10 +28,7 @@ use joltworks::{
         opening_proof::{ProverOpeningAccumulator, SumcheckId},
         rlc_polynomial::build_materialized_rlc,
     },
-    subprotocols::{
-        evaluation_reduction::{EvalReductionProof, EvalReductionProtocol},
-        sumcheck::SumcheckInstanceProof,
-    },
+    subprotocols::{evaluation_reduction::EvalReductionProof, sumcheck::SumcheckInstanceProof},
     transcripts::Transcript,
     utils::math::Math,
 };
@@ -91,7 +88,7 @@ impl<F: JoltField, T: Transcript, PCS: CommitmentScheme<Field = F>> ONNXProof<F,
         (poly_map, commitments)
     }
 
-    pub(super) fn output_claim(prover: &mut Prover<F, T>) {
+    pub(crate) fn output_claim(prover: &mut Prover<F, T>) {
         // Construct output polynomial
         let output_index = prover.preprocessing.model().outputs()[0];
         let output_computation_node = &prover.preprocessing.model()[output_index];
@@ -125,18 +122,16 @@ impl<F: JoltField, T: Transcript, PCS: CommitmentScheme<Field = F>> ONNXProof<F,
     /// Iterate over computation graph in reverse topological order
     /// Prove each operation using sum-check and virtual polynomials
     #[tracing::instrument(skip_all, name = "ONNXProof::iop")]
-    pub(super) fn iop(
+    pub(crate) fn iop(
         computation_nodes: &BTreeMap<usize, ComputationNode>,
         prover: &mut Prover<F, T>,
         proofs: &mut BTreeMap<ProofId, SumcheckInstanceProof<F, T>>,
         eval_reduction_proofs: &mut BTreeMap<usize, EvalReductionProof<F>>,
     ) {
         for (_, node) in computation_nodes.iter().rev() {
-            // Before each node is proven,
-            // perform eval reduction to reduce openings to a unique claim for the node output.
-            eval_reduction_proofs.insert(node.idx, EvalReductionProver::prove(prover, node));
-
-            proofs.extend(OperatorProver::prove(node, prover));
+            let (eval_reduction_proof, execution_proofs) = OperatorProver::prove(node, prover);
+            eval_reduction_proofs.insert(node.idx, eval_reduction_proof);
+            proofs.extend(execution_proofs);
         }
     }
 
@@ -234,197 +229,5 @@ impl<F: JoltField, T: Transcript, PCS: CommitmentScheme<Field = F>> ONNXProof<F,
             .values()
             .map(|poly| PCS::commit(poly, pcs).0)
             .collect()
-    }
-}
-
-struct EvalReductionProver;
-
-impl EvalReductionProver {
-    /// Reduce dual NodeOutput openings for a producer node before proving that node.
-    ///
-    /// This is the first integration step for PAZK 4.5.2 (2-to-1 only).
-    pub(super) fn prove<F: JoltField, T: Transcript>(
-        prover: &mut Prover<F, T>,
-        computation_node: &ComputationNode,
-    ) -> EvalReductionProof<F> {
-        let node_idx = computation_node.idx;
-        let openings = prover.accumulator.get_node_openings(node_idx);
-
-        let LayerData {
-            operands: _,
-            output,
-        } = Trace::layer_data(&prover.trace, computation_node);
-        let output_mle = MultilinearPolynomial::from(output.padded_next_power_of_two());
-
-        let (proof, reduced) =
-            EvalReductionProtocol::prove(&openings, output_mle, &mut prover.transcript)
-                .expect("Proving evaluation reduction should not fail");
-
-        prover
-            .accumulator
-            .reduced_evaluations
-            .insert(node_idx, reduced);
-
-        proof
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::onnx_proof::{AtlasSharedPreprocessing, HyperKZG};
-    use ark_bn254::{Bn254, Fr};
-    use atlas_onnx_tracer::{
-        model::{test::ModelBuilder, Model},
-        ops::{Add, Operator},
-        tensor::Tensor,
-    };
-    use common::VirtualPolynomial;
-    use joltworks::{
-        poly::{
-            multilinear_polynomial::{MultilinearPolynomial, PolynomialEvaluation},
-            opening_proof::{OpeningPoint, SumcheckId},
-        },
-        transcripts::Blake2bTranscript,
-    };
-    use rand::{rngs::StdRng, SeedableRng};
-    use std::collections::BTreeMap;
-
-    // All node output is used 1 to 2 times.
-    fn one_sub_model(rng: &mut StdRng, t: usize) -> Model {
-        let mut b = ModelBuilder::new();
-        let i = b.input(vec![t]);
-        let c = b.constant(Tensor::random_small(rng, &[t]));
-        let z = b.sub(i, c);
-        let out = b.sub(z, c);
-        b.mark_output(out);
-        b.build()
-    }
-
-    #[test]
-    fn eval_reduction_adds_reduced_evaluations() {
-        let n = 1 << 3; // 8 points, i.e. 3 variables
-        let output: Tensor<i32> = Tensor::construct((0..n).map(|n| n as i32).collect(), vec![n]);
-        let model = Model::default();
-        let producer_idx = 0;
-        let consumer1 = 1;
-        let consumer2 = 2;
-
-        let preprocessing = AtlasSharedPreprocessing::preprocess(model);
-        let mut trace = preprocessing.model().trace(&[]);
-        trace.node_outputs.insert(0, output.clone());
-        let mut prover = Prover::<Fr, Blake2bTranscript>::new(preprocessing, trace);
-
-        let producer_node = ComputationNode::new(0, Operator::Add(Add), vec![], vec![n]);
-
-        let output_mle = MultilinearPolynomial::from(output.padded_next_power_of_two());
-        let num_vars = output_mle.get_num_vars();
-
-        let point1 = prover.transcript.challenge_vector_optimized::<Fr>(num_vars);
-        let point2 = prover.transcript.challenge_vector_optimized::<Fr>(num_vars);
-        let claim1 = output_mle.evaluate(&point1);
-        let claim2 = output_mle.evaluate(&point2);
-
-        prover.accumulator.append_virtual(
-            &mut prover.transcript,
-            VirtualPolynomial::NodeOutput(producer_idx),
-            SumcheckId::NodeExecution(consumer1),
-            OpeningPoint::new(point1),
-            claim1,
-        );
-        prover.accumulator.append_virtual(
-            &mut prover.transcript,
-            VirtualPolynomial::NodeOutput(producer_idx),
-            SumcheckId::NodeExecution(consumer2),
-            OpeningPoint::new(point2),
-            claim2,
-        );
-
-        let _ = EvalReductionProver::prove(&mut prover, &producer_node);
-
-        let entries: Vec<_> = prover.accumulator.get_node_openings(producer_idx);
-
-        // We only seed two per-consumer openings and run eval-reduction proving.
-        assert_eq!(entries.len(), 2);
-        assert!(prover
-            .accumulator
-            .reduced_evaluations
-            .contains_key(&producer_idx));
-    }
-
-    #[test]
-    fn eval_reduction_single_opening_is_propagated() {
-        let n = 1 << 3; // 8 points, i.e. 3 variables
-        let output: Tensor<i32> = Tensor::construct((0..n).map(|n| n as i32).collect(), vec![n]);
-        let model = Model::default();
-        let producer_idx = 0;
-        let consumer = 1;
-
-        let preprocessing = AtlasSharedPreprocessing::preprocess(model);
-        let mut trace = preprocessing.model().trace(&[]);
-        trace.node_outputs.insert(producer_idx, output.clone());
-        let mut prover = Prover::<Fr, Blake2bTranscript>::new(preprocessing, trace);
-
-        let producer_node = ComputationNode::new(producer_idx, Operator::Add(Add), vec![], vec![n]);
-
-        let output_mle = MultilinearPolynomial::from(output.padded_next_power_of_two());
-        let num_vars = output_mle.get_num_vars();
-
-        let point = prover.transcript.challenge_vector_optimized::<Fr>(num_vars);
-        let claim = output_mle.evaluate(&point);
-
-        prover.accumulator.append_virtual(
-            &mut prover.transcript,
-            VirtualPolynomial::NodeOutput(producer_idx),
-            SumcheckId::NodeExecution(consumer),
-            OpeningPoint::new(point.clone()),
-            claim,
-        );
-
-        let _ = EvalReductionProver::prove(&mut prover, &producer_node);
-
-        let reduced = prover
-            .accumulator
-            .reduced_evaluations
-            .get(&producer_idx)
-            .expect("single opening should be propagated to reduced_evaluations");
-
-        assert_eq!(
-            reduced.r,
-            point.iter().map(|&c| c.into()).collect::<Vec<_>>()
-        );
-        assert_eq!(reduced.claim, claim);
-    }
-
-    #[test]
-    fn iop_runs_in_isolation_after_output_claim() {
-        let t = 1 << 4;
-        let mut rng = StdRng::seed_from_u64(0x1010);
-        let input = Tensor::<i32>::random_small(&mut rng, &[t]);
-        let model = one_sub_model(&mut rng, t);
-
-        let preprocessing = AtlasSharedPreprocessing::preprocess(model);
-        let trace = preprocessing.model().trace(&[input]);
-        let mut prover = Prover::<Fr, Blake2bTranscript>::new(preprocessing, trace);
-
-        let mut proofs: BTreeMap<ProofId, SumcheckInstanceProof<Fr, Blake2bTranscript>> =
-            BTreeMap::new();
-        let mut eval_reduction_proofs = BTreeMap::new();
-
-        ONNXProof::<Fr, Blake2bTranscript, HyperKZG<Bn254>>::output_claim(&mut prover);
-
-        let output_idx = prover.preprocessing.model().outputs()[0];
-        assert_eq!(prover.accumulator.get_node_openings(output_idx).len(), 1);
-        let computation_nodes = prover.preprocessing.model().nodes().clone();
-
-        ONNXProof::<Fr, Blake2bTranscript, HyperKZG<Bn254>>::iop(
-            &computation_nodes,
-            &mut prover,
-            &mut proofs,
-            &mut eval_reduction_proofs,
-        );
-
-        assert!(!proofs.is_empty());
-        assert!(eval_reduction_proofs.contains_key(&output_idx));
     }
 }
